nodetype changes - recreate old state
